Create OEM install script and playbook
This OEM build script in scripts/ ensures that necessary packages are
installed through apt before attempting to run Ansible. After that, it
runs the playbook for the OEM role installing and removing packages as
necessary and preparing some .desktop files for users to click when they
first sign in. Once the script completes, the VM should be in a state
that is nearly ready to ship to users.