Filezilla: Make sure we clarify the server is stu (#96)
It may not be clear to some people who only hear the server referred to
as "stu" that the server we have preconfigured is the server that they
want. Specify the name in parenthesis so that it's clear. Having just
"stu" feels like it also might be ambigious to some students who want
extra confirmation they're connecting to the right thing. Both together
seems like a decent medium.