Add Settings and About dialogs
Add a settings dialog so that users may specify the branch and URL to
use without having to manually modify a file in their home directory.
Additionally, provide an About dialog with information about the
application. This provides two benefits:
 1. The user knows where the program came from and provides a link to
    our GitHub
 2. There's more than just one thing in the toolbar

This also disables all the checkboxes when the Ansible script is
running, which is a preliminary resolution to #109, but not a full
enough solution to close that issue.

Further, there is now a try-except around the Vte spawn call because if
gksudo did not exist on the user's computer nothing at all would happen
except for a message being logged or printed to the console. The buttons
never got re-enabled because no call was ever made to
sub_command_exited. Now the error gets logged and that function is
called manually so a message can be displayed to the user.

Closes #108