Switch from gksudo to pkexec
gksudo has been deprecated and will no longer be shipping on many
distros. pkexec is the preferred replacement. This adds the necessary
items for us to use pkexec instead of gksudo.

We need a policykit policy. The updated task-shortcuts role installs it.
Additionally, the Python script is updated to call pkexec and support
its exit codes.