Remember user's chosen roles
Each run should make an attempt at saving the chosen roles so that they
can be read in as part of the config on the next run. This is easiest by
changing the configuration to JSON. This will invalidate all existing
configuration files; however, they will be fixed immediately at the end
of the next run of the script.

The only users who have probably modified their configuration files are
users who are developing the VM (since the in-app settings didn't exist
as of the Sylvia release). A post can be made in the virtual-machine team
announcing this change after this is merged.