Add Shellcheck linting
Ensure that we lint scripts/oem-build and some other known shell scripts
throughout the repository. Unfortunately, there is a nonzero amount of
shell code that gets generated using things like blockinfile that is
hard to actually lint this way. Two obvious files that we have that are
shell scripts are the oem-build script and the csvmprofile.

The csvmprofile is empty, as of this change. It gets content added to it
dynamically later but barring a full run & lint of files that we create,
at least making sure that what we track in the repo is valid is a good
start.