Add script to automate packer testing
This script will, given a PR ID and a template file, automatically
generate a packer var file and invoke packer to test a PR. It is
important to note that it is still a requirement that the code actually
be pushed to the branch (testing changes that are only local work just
as ineffectively as before) but at least it's (possibly) moderately
easier to test than with manually fetching the git info.

As an additional helper, since testing may happen on more platforms than
traditional building of the VM does, an attempt is made to guess the
proper audio device type for the host OS. If it doesn't work, then we
just give up and either default to "null" or prompt the user (if we're
running in interactive mode).

Some example invocations:

```
../scripts/packer_pr_test.py --var-file beta-vars.json 430 mint-build.json
```

or

```
../scripts/packer_pr_test.py 430 ubuntu-build.json
```