Refactor oem-build script (#430)
* Attempt automatically running as root

* Remove unnecessary --assume-yes

This script gets run with the noninteractive frontend specified from
within the packer scripts and if someone is actually running it in an
interactive situation, then we probably want to be interactive.

* Cleanup cd logic

A mix of `dirname` and `cd ..` is just confusing. Compute the root
directory of the project repo once and reuse it.

* Avoid hardcoding guest additions paths

Try to automatically detect the home directory of the user that the
script is running under. This avoids hardcoding the /home/oem path,
allowing this to run in other contexts or when the username has been
changed.

* Attempt cd early in oem-build

If there are issues switching to this directory, all kinds of wonky
things could happen and there's not much use in running the rest of the
script anyway if we won't be able to run the playbook.