Use loop syntax for getting users
Ansible 2.5 introduced the `loop` key which seems to be prefered now.
`with_dict` still works but we had a comment to clean this up so we may
as well. Reformatting the lookup also makes it a bit easier to tell
what's going on in this block. Since we're folding and chomping (`>-`)
then it shouldn't really matter.