Always use a string for mode
Using a leading 0 causes the value to get parsed as an octal integer,
per the YAML 1.1 spec (fwiw YAML 1.2 fixes this, but Ansible doesn't use
that). While in all our use cases this has been fine since we're using
fairly typical modes, it would get akward if we needed to set one of the
sticky/setuid/setgid bits (2755, etc) since that's no longer octal. We'd
still need a leading zero or to quote it for consistency. Since
directory_mode is already a little wonky, it's likely best to just
always quote the mode.