Update hashlint to support new hash structure
This automatically uses the 'hash' to determine the values that should
be used for `{{ ansible_architecture }}` and uses `jinja2` to perform
the interpolation, just as Ansible would do.

The assumption is that if `hash` is a dictionary, it is a mapping of
`ansible_architecture` values to the expected hash and that if there is
not a dictionary that there is a platform-agnostic download and hash. If
it turns out we need more complicated processing for the URLs we'll have
to rethink some of this logic, but this is smart enough for now without
being too smart.

Now that this adds another two (slower) items to the list of things to
download, we make use of the `aiohttp` library to download the items in
parallel.