Cleanup y86ref.1
This resolves the following minor issues with the man page:
 - options have been reordered to match program help text ordering
 - whitespace at the end of files has been removed (mandoc STYLE issue)
 - date format has been changed (and the date has been updated (mandoc
   WARNING issue)

This selected date format is based on `man-pages(7)`'s recommendation;
however a cursory look at the existing mandb
`zcat /usr/share/man/man1/*.1.gz | grep -i '^\.Th'` shows that there is
very little consistency in how dates are formatted. It seems that as far
as `mandoc(1)` is concerned, the two valid formats are "Month DD, YYYY"
or "YYYY-MM-DD". The latter feels less necessary to have to worry about
localization with.