Improve /etc/os-release parsing
Update the parsing to more closely following the recommendations from
systemd. os-release(5) specifies that the file will have all strings in
UTF-8 format, that we should prefer the file in /etc/ but also check
/usr/lib if /etc/ doesn't exist, and that we should accept quoted
values.

Per the recommendations in os-release(5) or from the upstream systemd
project, the Python code has been refactored to use the ast module to
perform the literal evaluation. The code is not used exactly so as to
adjust for our use case.