Remove unnecessary block
We actually always want to ensure that the `finch.install_path`
directory exists and now that we're using `checksum` on `get_url` that
won't execute if the hash matches. We can then selectively unarchive the
`zip` only when the download changed something. Even still, we may want
to eventually consider always unarchiving.