Add action to automatically label pull requests
We have had labels for the various courses for awhile; however, they
have gone unused in the vast majority of cases. GitHub has an action for
labeling PRs based on files modified. It requires a configuration file
mapping labels to paths which would be pretty duplicative in this repo
since we already have the playbook that performs that mapping.

A fairly simply Python script is added to handle converting the Ansible
playbook file to a Labeler configuration file. If we wanted to handle
additional labels in the future then some refactoring would be required;
however, this is a decent start.

Additionally, this is a separate workflow from the rest of ours. Since
modifying pull request labels require write access to the repository and
(nearly) all of our PRs originate from forks, then the `pull_request`
event's default token will not be sufficient here. Instead,
`pull_request_target` will be used. This means that we will always use
the playbook file and script on the default branch. This means that if
`local.yml` is modified to add new courses or tags, the PR will still be
labeled according to the config on the main branch. Additionally, it
prevents malicious modifications to the label config generator script
from having write access to the repository since it will use the version
on the main branch.

Finally, `pull_request_target` events are only triggered when the
workflow is already on the default branch; therefore, we will not see
the effects of this PR until it has been merged.