Fix wrapper linting and formatting
pylint was grumpy due to a use of %-style formatting instead of
f-strings. Since Mint has a new enough version of Python to support
f-strings and since they're nicer, this adopts the recommendation of the
linter. Note that standard Python convention is to still use the %-style
formatting (but not the % operator) in calls to the logging library so
that formatting can happen lazily.

I've also run `black` on the code base to improve the style and
formatting conventions. Most of these are related to spacing, line
breaks, and the particular quote character in use.