Migrate to XDG Dirs and YAML
Linux Mint now contains pyxdg (as python3-xdg) and PyYAML by default on
a base install. We can use both of these to make parts of the app a bit
nicer. First, we can easily comply with the XDG Base Directories
standard which we weren't really doing before by writing directly into
~/.config and ~/.cache without putting things in our application's
directory. Migrating to these new directories is a breaking change in
itself. At the same time, we may as well move to YAML for the config
file since basically everything else we touch is YAML anyway.
Theoretically, if we really took care we could migrate a user config
from the old format to the new but I am going to argue against adding
code for such a migration since almost no users upgrade their VM from
one Mint release to the next and since this will likely first appear in
a spring VM, adding code for legacy migrations does not seem valuable.
Any users who dangerously ride along on `main` are probably the author
or reviewer of this PR.