Convert Packer definition to HCL2
This rewrites the packer JSON files to use the HCL2 language support
added in 1.5 and stabilized in 1.7. The initial conversion was performed
using the `hcl2_upgrade` command but quite a lot of custom work has been
done on top of that.

A base shared definition for the VM is created that defines all the
shared options (CPU, memory, SSH stuff, etc) and then that is
specialized in the `build` block for each `source`, this is where the
differences between Ubuntu and Mint are specified.

Additionally, we now store version information as complex objects that
meet our needs a bit more closely. This is required because some of the
interpolation that we did in JSON is no longer available.

As I was reviewing the VirtualBox builder config I also moved a few
things from manual `VBoxManage` commands to actual packer definitions
but some of those could have been done in the JSON format as well.