Fix the Ubuntu update workflow to work with HCL
The change to HCL broke the update workflow because it's not JSON and
the JSON file we were parsing was deleted. This uses an
`auto.pkrvars.hcl` file to set the Ubuntu version (for which an updated
version is intentionally not committed so that the workflow will have
work to do). This means that check will fail until the workflow runs and
opens its own PR (becuse validation fails if the checksum can't be
found).

This is done _in addition_ to `version` so that the the friendlier
version can be used in contexts where the patch version doesn't really
matter (like the parent directory or the build metadata).