Explicitly use Ubuntu 22.04 in workflows
We had been using 20.04 and while the 22.04 runner is now GA, it is not
yet the default for `ubuntu-latest`. And even if it were, we were still
hardcoding 20.04 anyway. Because we've updated the version of Mint and
of Ubuntu used within the config, we should update the workflows to
match.