Fix linting syntax error in test files
This changes the role names to match the underlying role's actual name.
Orginally, in #422, we renamed the roles to use underscores because that
is what `ansible-lint` wanted. Unfortunately, at the time,
`ansible-lint` did not have opinions on the content of the `test/`
directories (where we failed to update the names to match) and so this
has gone undetected as an issue for quite some time (as we do not ever
run any tests). This check seems to have been added in `ansible-lint`
somewhere around v6.13