Add `names` to all `test.yml` plays
The `test.yml` files seem to now be validated against the `ansible-lint`
which requires all plays to have a `name`. This was not done by the
default `ansible-galaxy init` that was used to setup the roles (quite a
while ago now). This adds the required field for basic validation to
work.