Update to Mint 21.3 Virginia
This updates to Mint 21.3. As part of the release, Mint changed to using
syslinux on BIOS installs. I really don't want to learn how that works
for the boot command so instead, this migrates to UEFI which Mint
promises is better supported in 21.3 and which VirtualBox continually is
improving support for. The boot command gets some minor tweaks but these
match what was planned for QEMU on AARCH64 anyway.