Use different name for builder and exported VM
This allows building a new VM for the current semester while a previous
build of the same semester is still registered in VirtualBox. It doesn't
help having multiple builds of the same VM going at the same time but it
makes things slightly smoother. The exported VM gets to keep the same
name pattern it's always had via the `vmname` argument to
`VBoxManage export`.

Additionally, the `legacy_isotime` function invocation is replaced with
the recommended migration path.