Enable good USB
We previously started to allow XHCI with them being included in the base
VirtualBox and no longer restrictively licensed; let's go ahead and
actually turn them on all the way after we're done with the initial
build. Even though it's unlikely too many things will get mapped into
the VM, it's nice to have a sensible and good default.